mnt: refactor parachute implementation
