+++ docs/user/rocket/rocket_usage.rst
@@ -295,7 +295,9 @@ apogee and another that will be deployed at 800 meters above ground level:
 
 .. jupyter-execute::
 
-    main = calisto.add_parachute(
+    from rocketpy import HemisphericalParachute
+
+    main = HemisphericalParachute(
         name="Main",
         cd_s=10.0,
         trigger=800,
@@ -307,7 +309,7 @@ apogee and another that will be deployed at 800 meters above ground level:
         porosity=0.0432,
     )
 
-    drogue = calisto.add_parachute(
+    drogue = HemisphericalParachute(
         name="Drogue",
         cd_s=1.0,
         trigger="apogee",
@@ -318,6 +320,8 @@ apogee and another that will be deployed at 800 meters above ground level:
         height=1.5,
         porosity=0.0432,
     )
+    calisto.add_parachute(main)
+    calisto.add_parachute(drogue)
 
 .. seealso::
