+++ rocketpy/__init__.py
@@ -36,9 +36,9 @@ from .rocket import (
     FreeFormFin,
     FreeFormFins,
     GenericSurface,
+    HemisphericalParachute,
     LinearGenericSurface,
     NoseCone,
-    Parachute,
     PointMassRocket,
     RailButtons,
     Rocket,
