+++ rocketpy/rocket/parachutes/__init__.py
@@ -1,0 +1,1 @@
+from .hemispherical_parachute import HemisphericalParachute
